+++ docs/release_notes.md
@@ -4,6 +4,7 @@
 * __移除CLI工具__
 * __停止对PyPy的官方支持__
 * __移除参数`is_half_char_fn`以及相对应的常量`DEFAULT_HALF_CHARS`__
+* __将参数`is_end_char_fn`替换为新参数`end_chars`，废除常量`DEFAULT_END_CHARS`。__
 
 ## v3.4.0 (2019-5-1)
 * __Pillow版本限制放宽至`>= 5, < 7`__